backend/requirements.txt
fastapi
uvicorn
pydantic 
google-genai
chromadb
pandas
datasets
sentence_transformers
transformers
accelerate